+++ CHANGELOG.md
@@ -8,6 +8,8 @@ All notable changes to this project will be documented in this file.
 
 - Description: Aero Hand Open (MJCF)
 - Description: Aero Hand Open (URDF)
+- Description: ToddlerBot 2XC (MJDF)
+- Description: ToddlerBot 2XM (MJDF)
 
 ### Changed